backend/requirements.txt
blinker==1.4
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
pyodbc==4.0.28
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.0
